+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.0] - 2023-05-28
 ☮️ Peace in the world, or the world in pieces. 🕊️
 ### Added
 - New translation strings, Radiation Centre is now fully localised.